requirements.txt
altair==6.0.0
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lightgbm==4.6.0
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
protobuf==6.33.6
pyarrow==23.0.1
pydeck==0.9.1
PyJWT==2.12.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
xgboost==3.1.1
playwright==1.58.0